refactored original sample and removed camera view + BBoxs
