--- lib/ui/camera_view_singleton.dart
@@ -1,10 +1,0 @@
-import 'dart:ui';
-
-/// Singleton to record size related data
-class CameraViewSingleton {
-  static double ratio;
-  static Size screenSize;
-  static Size inputImageSize;
-  static Size get actualPreviewSize =>
-      Size(screenSize.width, screenSize.width * ratio);
-}
