+++ lib/ui/home_view.dart
@@ -1,13 +1,16 @@
-import 'package:flutter/cupertino.dart';
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 import 'package:object_detection/ml_models.dart';
 import 'package:object_detection/tflite/recognition.dart';
 import 'package:object_detection/tflite/stats.dart';
-import 'package:object_detection/ui/box_widget.dart';
-import 'package:object_detection/ui/camera_view_singleton.dart';
 
 import 'package:object_detection/theme.dart';
-import 'camera_view.dart';
+import 'package:object_detection/ui/bottom_sheet.dart';
+import 'live_camera_preview.dart';
+import 'stats_row.dart';
+
+
 
 /// [HomeView] stacks [CameraView] and [BoxWidget]s with bottom sheet for stats
 class HomeView extends StatefulWidget {
@@ -19,15 +22,23 @@ class _HomeViewState extends State<HomeView> {
   /// Results to draw bounding boxes
   List<Recognition> results;
 
-  /// Realtime stats
-  Stats stats;
-
   /// Scaffold Key
   GlobalKey<ScaffoldState> scaffoldKey = GlobalKey();
 
   bool statsOpen = false;
 
   MLModels mlModel = MLModels.YOLOV5;
+
+  CameraController cameraController;
+
+  @override
+  void initState() {
+    cameraController = CameraController(
+      GetIt.I<List<CameraDescription>>()[0],
+      ResolutionPreset.low,
+    );
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -63,121 +74,34 @@ class _HomeViewState extends State<HomeView> {
             child: CameraView(resultsCallback, statsCallback)
           ),
 
-          // Bounding boxes
-          Align(
-            alignment: Alignment.topCenter,
-            child: boundingBoxes(results)
+          /*
+          // change camera button
+          Positioned(
+            bottom: 50,
+            right: 50,
+            child: Container(
+              height: 100,
+              child: IconButton(
+                icon: Icon(Icons.cameraswitch),
+                onPressed: () {},
+              ),
+            ),
           ),
+          */
 
           // Bottom Sheet
-          stats != null
-            ? Positioned(
-              top: statsOpen ? null : MediaQuery.of(context).size.height - 50,
-              bottom: statsOpen ? 0 : null,
-              width: MediaQuery.of(context).size.width,
-              child: Container(
-                decoration: BoxDecoration(
-                  color: dcaitiBlue,
-                  borderRadius: BorderRadius.only(topLeft: Radius.circular(8), topRight: Radius.circular(8))
-                ),
-                child: Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      StatsRow('Inference time:',
-                          '${stats.inferenceTime} ms'),
-                      StatsRow('Total prediction time:',
-                          '${stats.totalElapsedTime} ms'),
-                      StatsRow('Pre-processing time:',
-                          '${stats.preProcessingTime} ms'),
-                      StatsRow('Frame',
-                          '${CameraViewSingleton.inputImageSize?.width} X ${CameraViewSingleton.inputImageSize?.height}'),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [MLModels.YOLOV5, MLModels.YOLOV7, MLModels.CascadeRCNN].map((e) => 
-                          Row(
-                            children: 
-                            [
-                              Radio<MLModels>(
-                                value: e,
-                                groupValue: mlModel,
-                                activeColor: dcaitiGreen,
-                                onChanged: (MLModels value) {
-                                  setState(() {
-                                    mlModel = value;
-                                  });
-                                },
-                              ),
-                              Text(e.name),
-                            ]
-                          ,)
-                        ).toList()                 
-                      ),
-                      Text(
-                        "\n\nThis project is a research coorporation between DCAITI and the TU Berlin.\n"
-                        "Developers: Dario Klepoch, Marvin Beese, Clemens Lotthermoser",
-                        textScaleFactor: 0.8,
-                      )
-                    ],
-                  ),
-                ),
-              ),
+          Positioned(
+            top: statsOpen ? null : MediaQuery.of(context).size.height,
+            bottom: statsOpen ? 0 : null,
+            width: MediaQuery.of(context).size.width,
+            child: BottomInfoSheet(
+              mlModel: mlModel,
+              onChanged: (value) => setState(() {
+                mlModel = value;
+              }),
             )
-          : SizedBox()
+          )
         ],
-      ),
-    );
-  }
-
-  /// Returns Stack of bounding boxes
-  Widget boundingBoxes(List<Recognition> results) {
-    if (results == null) {
-      return Container();
-    }
-    return Stack(
-      children: results
-          .map((e) => BoxWidget(
-                result: e,
-              ))
-          .toList(),
-    );
-  }
-
-  /// Callback to get inference results from [CameraView]
-  void resultsCallback(List<Recognition> results) {
-    setState(() {
-      this.results = results;
-    });
-  }
-
-  /// Callback to get inference stats from [CameraView]
-  void statsCallback(Stats stats) {
-    setState(() {
-      this.stats = stats;
-    });
-  }
-
-  static const BOTTOM_SHEET_RADIUS = Radius.circular(24.0);
-  static const BORDER_RADIUS_BOTTOM_SHEET = BorderRadius.only(
-      topLeft: BOTTOM_SHEET_RADIUS, topRight: BOTTOM_SHEET_RADIUS);
-}
-
-/// Row for one Stats field
-class StatsRow extends StatelessWidget {
-  final String left;
-  final String right;
-
-  StatsRow(this.left, this.right);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 8.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [Text(left), Text(right)],
       ),
     );
   }
