+++ lib/ui/stats_row.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+
+
+/// Row for one Stats field
+class StatsRow extends StatelessWidget {
+  final String left;
+  final String right;
+
+  StatsRow(this.left, this.right);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8.0),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [Text(left), Text(right)],
+      ),
+    );
+  }
+}
